+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/ProjectUpgradesPage/UpgradeWizard/steps/FundingCycleStep/LaunchFundingCycleForm.tsx
@@ -3,7 +3,7 @@ import { Button, Divider, Form, Input, Space } from 'antd'
 import { useIsNftProject } from 'components/Create/hooks/DeployProject/hooks'
 import UnsavedChangesModal from 'components/modals/UnsavedChangesModal'
 import RichButton from 'components/RichButton'
-import FundingDrawer from 'components/v2v3/shared/FundingCycleConfigurationDrawers/FundingDrawer'
+import { FundingDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/FundingDrawer'
 import NftDrawer from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer'
 import { RulesDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/RulesDrawer'
 import { TokenDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/TokenDrawer'
