+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/FundingDrawer/FundingForm/DistributionSplitsSection/PayoutConfigurationExplainerCollapse.tsx
